refactor(state): state snapshot use newtype pattern
refactor(state): state snapshot use newtype pattern
