+++ backend/nyanpasu-core/src/state/manager/persistent_builder.rs
@@ -6,7 +6,10 @@ use fs_err::tokio as fs;
 use serde::{Serialize, de::DeserializeOwned};
 use std::{io::Write, sync::Arc};
 
-use super::{super::error::*, *};
+use super::{
+    super::{StateSnapshot, error::*},
+    *,
+};
 
 use crate::format::{Format, YamlFormat};
