+++ backend/nyanpasu-core/src/state/mod.rs
@@ -3,6 +3,7 @@ pub mod builder;
 pub mod coordinator;
 pub mod error;
 pub mod manager;
+mod snapshot;
 
 #[cfg(test)]
 mod tests;
@@ -11,3 +12,4 @@ pub use ack::*;
 pub use builder::*;
 pub use coordinator::*;
 pub use manager::*;
+pub use snapshot::*;
